Add confirmation before deleting logs

+++ out/explorer.js
@@ -113,35 +113,42 @@ function downloadLog(logId, context) {
 }
 exports.downloadLog = downloadLog;
 function deleteLogs() {
-    vscode.window.withProgress({
-        location: vscode.ProgressLocation.Notification,
-        title: "Submitting delete job...",
-        cancellable: true,
-    }, (progress, token) => __awaiter(this, void 0, void 0, function* () {
-        return new Promise((resolve, reject) => {
-            apexlog.sfdx
-                .command("force:data:soql:query", ["-q", "SELECT Id FROM ApexLog", "-r", "csv"], false)
-                .then((response) => {
-                fs.writeFileSync(apexlog.config.getWorkspaceFolder() + "/out.csv", response);
+    return __awaiter(this, void 0, void 0, function* () {
+        const confim = yield vscode.window.showWarningMessage("Are you sure you want to delete logs? This will delete all of the logs in your org.", { modal: true }, "Yes");
+        if (confim !== "Yes")
+            return;
+        vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: "Async Log Delete Job",
+            cancellable: false,
+        }, (progress, token) => __awaiter(this, void 0, void 0, function* () {
+            return new Promise((resolve, reject) => {
+                progress.report({ message: "Querying for logs to delete..." });
                 apexlog.sfdx
-                    .command("force:data:bulk:delete", ["-s", "ApexLog", "-f", "out.csv"])
-                    .then((result) => {
-                    vscode.window.showInformationMessage("Delete job submitted successfully");
-                    resolve(result);
-                })
-                    .catch((error) => {
-                    if (error.message === "Unable to find any data to create batch") {
-                        vscode.window.showInformationMessage("Unable to find any log data to delete.");
-                        resolve();
-                    }
-                    else {
-                        vscode.window.showErrorMessage("Unable to create delete job: " + error.message);
-                        resolve();
-                    }
+                    .command("force:data:soql:query", ["-q", "SELECT Id FROM ApexLog", "-r", "csv"], false)
+                    .then((response) => {
+                    fs.writeFileSync(apexlog.config.getWorkspaceFolder() + "/out.csv", response);
+                    progress.report({ message: "Requesting bulk deletion of logs..." });
+                    apexlog.sfdx
+                        .command("force:data:bulk:delete", ["-s", "ApexLog", "-f", "out.csv"])
+                        .then((result) => {
+                        vscode.window.showInformationMessage("Request to delete logs sent successfully, it may take a few moments for your logs to be deleted.");
+                        resolve(result);
+                    })
+                        .catch((error) => {
+                        if (error.message === "Unable to find any data to create batch") {
+                            vscode.window.showInformationMessage("Logs are already empty, nothing to delete.");
+                            resolve();
+                        }
+                        else {
+                            vscode.window.showErrorMessage("Unable to create delete job: " + error.message);
+                            resolve();
+                        }
+                    });
                 });
             });
-        });
-    }));
+        }));
+    });
 }
 exports.deleteLogs = deleteLogs;
 function deleteTraceFlag(context) {
